+++ lib/commands/impl/install/install_get.dart
@@ -1,7 +1,7 @@
 import '../../../common/utils/pubspec/pubspec_utils.dart';
 
 Future<void> installGet([bool runPubGet = false]) async {
-  if (PubspecUtilsExt.useSmGetx) return;
+  if (PubspecUtilsExt.getPackagePrefix != 'get') return;
 
   PubspecUtils.removeDependencies('get', logger: false);
   // TODO: 当 get 5 正式发布之后修改